May tigtry lang ako. Yun tab system chuchu
