+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "cmake.sourceDirectory": "D:/Files/Da Documents/Programming/Git/foundcalc/foundcalc/windows"
+}
