+++ foundcalc/lib/pages/design_page.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+
+class DesignPage extends StatefulWidget {
+  final String title;
+
+  DesignPage({required this.title});
+
+  @override
+  _DesignPageState createState() => _DesignPageState();
+}
+
+class _DesignPageState extends State<DesignPage> {
+  // You can add state variables and methods here
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Color(0xFF212121),
+      appBar: AppBar(
+        title: Text(
+          widget.title,
+          style: TextStyle(color: Colors.white)
+        ),
+        backgroundColor: Colors.black,
+      ),
+      body: Center(
+        child: Text(
+          'Details for ${widget.title}',
+          style: TextStyle(color: Colors.white, fontSize: 24),
+        ),
+      ),
+    );
+  }
+}
